Exercise_10/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-groq>=0.1.0
chromadb>=0.5.0
fastembed>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.31.0
pypdf>=4.0.0
